+++ CHANGES.rst
@@ -28,7 +28,9 @@ Changes
 - Connection can be destroyed before response get processed if
   `await aiohttp.request(..)` is used #1981
 
-- MultipartReader doesn't work with -OO #1969
+- MultipartReader does not work with -OO #1969
+
+- Fixed `ClientPayloadError` with blank `Content-Encoding` header #1931
 
 
 2.1.0 (2017-05-26)